+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ### To be Released
 
+### 1.21.0
+
 * fix(backup-download): Now download the last successful backup [#663](https://github.com/Scalingo/cli/pull/663)
-* fix(review-app-url): add app url in the tablewriter [#664](https://github.com/Scalingo/cli/pull/664)
+* fix(review-app-url): add app url in the table [#664](https://github.com/Scalingo/cli/pull/664)
 * fix(backups-download): print error on stderr if the backup status is not 'done' [#665](https://github.com/Scalingo/cli/pull/665)
 * feat(command): add command `cron-tasks` to list cron tasks of an application [#670](https://github.com/Scalingo/cli/pull/670)
 * fix(run): allow an equals sign (=) in environment variable values [#674](https://github.com/Scalingo/cli/pull/674)